+++ source/gear/corrosion-shattering.js
@@ -37,6 +37,6 @@ module.exports = new GearTemplate("Shattering Corrosion",
 	}
 ).setTargetingTags({ type: "single", team: "foe", needsLivingTargets: true })
 	.setSidegrades("Flanking Corrosion", "Harmful Corrosion")
-	.setModifiers({ name: "Power Down", stacks: 40 }, { name: "Frail", stacks: 4 })
+	.setModifiers({ name: "Power Down", stacks: 20 }, { name: "Frail", stacks: 4 })
 	.setStagger(2)
 	.setDurability(15);