+++ source/items/protectionpotion.js
@@ -2,13 +2,13 @@ const { ItemTemplate } = require("../classes");
 const { selectAllAllies } = require("../shared/actionComponents");
 
 module.exports = new ItemTemplate("Protection Potion",
-	"Adds 50 block to all allies",
+	"Adds 50 protection to all allies",
 	"Untyped",
 	30,
 	selectAllAllies,
 	false,
 	([target], user, isCrit, adventure) => {
-		target.block += 50;
+		target.protection += 50;
 		return `${user.getName(adventure.room.enemyIdMap)} gains protection.`;
 	}
 );